+++ NEWS.md
@@ -1,3 +1,8 @@
+# boxoffice 1.2.0
+
+* Adds ability to scrape top grossing movies - based on domestic (American), international, and total gross.
+* Changes some tests as the website's top grossing page has been updated.
+
 # boxoffice 1.1.0
 
 * Adds message about boxofficemojo.com's scraping rules